fixes #13104

This PR adds microarchitecture specific optimization flags to Clang / LLVM starting from version 3.9:
```console
$ spack -d install zlib %clang
[...]
==> [2019-10-17-20:21:19.885939] Successfully installed zlib
  Fetch: 0.00s.  Build: 4.10s.  Total: 4.10s.
[+] /home/culpo/PycharmProjects/spack/opt/spack/linux-ubuntu18.04-broadwell/clang-8.0.0/zlib-1.2.11-5i54k54rkpieapytrliahvgb52vcr2km
==> [2019-10-17-20:21:19.888574] WRITE LOCK: /home/culpo/PycharmProjects/spack/opt/spack/.spack-db/lock[0:0] [Acquiring]
==> [2019-10-17-20:21:19.888965] WRITE LOCK: /home/culpo/PycharmProjects/spack/opt/spack/.spack-db/lock[0:0] [Acquired]
==> [2019-10-17-20:21:19.903866] WRITE LOCK: /home/culpo/PycharmProjects/spack/opt/spack/.spack-db/lock[0:0] [Released]

$ head -n 4 spack-cc-zlib-5i54k54.out.log 
[cc] /usr/bin/clang-8 -march=x86-64 -mcpu=broadwell -c ztest7285.c
[cc] /usr/bin/clang-8 -march=x86-64 -mcpu=broadwell -c -fPIC -O2 ztest7285.c
[cc] /usr/bin/clang-8 -march=x86-64 -mcpu=broadwell -c -fPIC -O2 ztest7285.c
[cc] /usr/bin/clang-8 -march=x86-64 -mcpu=broadwell -w -c -fPIC -O2 -fPIC ztest7285.c
```
similarly to what was done for GCC in #3206.

For the time being mixed toolchains (e.g. with `clang`, `clang++` and `gfortran`) are disallowed, meaning no microarchitecture specific flags will be injected during compilation. A user warning is displayed in such cases:
```console
$ spack compiler info clang@3.9
clang@3.9.1:
        paths:
                cc = /usr/bin/clang-3.9
                cxx = /usr/bin/clang++-3.9
                f77 = /usr/bin/gfortran
                fc = /usr/bin/gfortran
        modules  = []
        operating system  = ubuntu18.04

$ spack install zlib %clang@3.9
==> Installing zlib
[...]
==> Warning: microarchitecture specific optimizations are not supported yet on mixed compiler toolchains [check clang@3.9.1 for further details]
==> Using cached archive: /home/culpo/PycharmProjects/spack/var/spack/cache/zlib/zlib-1.2.11.tar.gz
==> Staging archive: /tmp/culpo/spack-stage/spack-stage-zlib-1.2.11-xftam6nhhoqsw46m6f4xoyhfcdbd32jf/zlib-1.2.11.tar.gz
==> Created stage in /tmp/culpo/spack-stage/spack-stage-zlib-1.2.11-xftam6nhhoqsw46m6f4xoyhfcdbd32jf
==> No patches needed for zlib
==> Building zlib [Package]
==> Executing phase: 'install'
==> Successfully installed zlib
  Fetch: 0.01s.  Build: 4.99s.  Total: 4.99s.
[+] /home/culpo/PycharmProjects/spack/opt/spack/linux-ubuntu18.04-broadwell/clang-3.9/zlib-1.2.11-xftam6nhhoqsw46m6f4xoyhfcdbd32jf

```