@alalazo 2019-10-17
@adamjstewart Do you have some time to test drive this PR on a Mac and check if there are issues?