@adamjstewart 2019-10-17
@alalazo I can build a couple of packages if you want, but I can't test much on this branch since I use Clang + GFortran (scipy needs a fortran compiler).