@alalazo 2019-10-17
@adamjstewart It would be good to know if things continue to work as it is now for your case of mixed compilers and you just get a warning as expected. Also, if you use `flang` optimizations should kick-in.