@alalazo 2019-10-18
For later reference the error to be fixed in the tests is:
```console
Creating a mock package installation
==> Error: cannot produce optimized binary for micro-architecture 'ivybridge' with clang@9.1.0-apple [supported compiler versions are 3.9:]
```
I'll work on that asap. In the meanwhile I noticed 4 false positives in every unit test of every PR I sampled:
```console
Testing 'spack env deactivate'
'SPACK_ENV' is not set ... FAIL
[...]
Testing 'despacktivate'
'SPACK_ENV' is not set ... FAIL
[...]
'SPACK_OLD_PS1' is not set ... FAIL
SPACK_OLD_PS1 was set:
```
I'll try to have a look at that too.