@alalazo 2019-10-18
Should be ready for a test drive on a Mac. Apple's Clang will be demoted to build on `x86_64` for now.