@tgamblin 2019-10-19
Ok, I was able to build `py-scipy ^python@3.8.0` on `mojave` using this, and it works.

With and without this, there seems to be a patch that doesn't apply properly with Python 3.8, and when scipy is building, I get numerous system dialogs saying python quit unexpectedly.  If I dismiss them, things go along fine, but it's pretty annoying.  it doesn't seem to be related to this PR, though, so we can debug that separately.