@tgamblin 2019-10-19
i guess it's actually while installing `numpy` that I get the weird dialogs.  Curious if @adamjstewart gets the same thing.