@adamjstewart 2019-10-19
@tgamblin see #12356. Python 3.8.0 is still very buggy. I’m not using it anymore.