@adamjstewart 2019-10-21
@alalazo is this the expected behavior?
```console
$ spack spec zlib arch=darwin-catalina-ivybridge
Input spec
--------------------------------
zlib arch=darwin-catalina-ivybridge

Concretized
--------------------------------
==> Error: cannot produce optimized binary for micro-architecture 'ivybridge' with clang@11.0.0-apple [supported compiler versions are 3.9:]
```
11.0 > 3.9