@tgamblin 2019-10-21
@alalazo: a thought, after talking to @adamjstewart about this: can we still use the opt flags on the C compilers in this situation, and call the target the *optimized* name instead of going all the way back to `x86_64`?  Basically just remove opt flags for fortran.