@alalazo 2019-10-21
@adamjstewart  The behavior above is expected, but the error message is terrible. It should say that the suffixes don't match and we have no rule for the `apple` suffix.

@tgamblin Here the issue is not the opt flags for Fortran, but the fact that Apple has its own numbering. If somebody could add the corresponding `"versions": X.Y-apple:` in the JSON file everything would work as expected. I can do it if there's any document online, but otherwise I think I need acces to a Mac to start working on it.

Also, at the moment we have only one set of optimization flags valid for all the 4 languages (C, C++, F90, F77). I'll add an issue to separate them but that will be unrelated to the comments above I think. It will be needed to support optimization and mixed compiler toolchains though.