@adamjstewart 2019-10-21
Not so subtle plug for separating Clang and AppleClang into separate compilers: #12011 😏 

And yes, better mixed compiler support would be phenomenal. A lot of compilers like AppleClang and NAG don't come with the full C/C++/Fortran set, and have to be mixed with other compilers to work properly. For now, we've been assuming that the other compiler is GCC, which usually works, until you decide to use Flang or something else instead. This isn't exactly an uncommon use case, as every macOS user faces this same dilemma.